fix: address 6 P1 bugs — threading, views clipping, overflow, dispatch
- BUG-04: swap upper_bound/lower_bound in XYView/XYZView boundary search
- BUG-05: make plot() dispatch exhaustive with ValueError, stop swallowing exceptions
- BUG-06: use int64_t cast in _nanoseconds/_microseconds to prevent 2038 overflow
- BUG-07: add holds_alternative guard before std::get in _range_based_update
- THREAD-02: emit _state_changed() after unlocking m_mutex in set_data overloads
- THREAD-03: copy data snapshot under lock, run _resample_impl outside _data_mutex

+++ src/SciQLopTracer.cpp
@@ -27,12 +27,12 @@ namespace _impl
 {
 int _microseconds(const double epoch)
 {
-    return static_cast<int>((epoch * 1e3 - static_cast<int>(epoch * 1e3)) * 1e3);
+    return static_cast<int>((epoch * 1e3 - static_cast<int64_t>(epoch * 1e3)) * 1e3);
 }
 
 int _nanoseconds(const double epoch)
 {
-    return static_cast<int>((epoch - static_cast<int>(epoch)) * 1e9);
+    return static_cast<int>((epoch - static_cast<int64_t>(epoch)) * 1e9);
 }
 
 inline std::string _render_value(const double value)
